Julia's Max-Plus Algebra Toolbox

Julia's Max-Plus Algebra Toolbox

Main Max-Plus structure

Overriden operators

Max-Plus constants

Max-Plus Zeros

Max-Plus Ones

Max-Plus Identity Matrix

Max-Plus Convertion

Power

Max-Plus Matrix operations

Display control of Max-Plus scalar and Matrices

Index

Bibliography and references

Scilab and Julia function names

Max+ core functions

Dynamic linear maxplus system

Flowshop

Incompatibility

Julia's Max-Plus Algebra Toolbox

Prerequisite

Installation and your first Max-Plus code

Deeper dive with Julia's Max-Plus toolbox

Contribution

Max-Plus Tutorials

Max-Plus introduction with Julia

Manipulation of Max-Plus Dynamic Linear Systems with Julia

Flowshop

Main Max-Plus structure

Scalar

MaxPlus.MP — Type

MP

Immutable Julia structure for Max-Plus scalar. Promote a number of type Float64 or Int64 to a number in the tropical semi-ring (max, +) (ℝ ∪ {-∞}, ⨁, ⨂) where ℝ is the domain of reals, ⨁ is the usual multiplication and ⨂ is the usual maximum.

Notes

MP(3) is the equivalent to ScicosLab code: #(3)

Examples

julia> a = MP(3.5)
Max-Plus 3.5

julia> typeof(a)
MP

julia> a = MP(3)
Max-Plus 3

julia> typeof(a)
MP

Dense Matrix and Dense Vector

MaxPlus.MP — Method

MP(A::Array)

Convert a dense array from classic algebra to a Max-Plus dense array. Note: if the array alaready contains Max-Plus number then the MP() is not necessary since Max-Plus number contaminate other numbers.

Examples (constructor)

julia> A = MP([1.0 -Inf; 0.0 4])
2×2 Max-Plus dense matrix:
  1.0   -Inf
  0.0    4.0

julia> typeof(A)
Matrix{MP} (alias for Array{MP, 2})

Examples (constructor not needed)

julia> A = [MP(1.0) -Inf; 0.0 4]
2×2 Max-Plus dense matrix:
  1.0   -Inf
  0.0    4.0

julia> typeof(A)
Matrix{MP} (alias for Array{MP, 2})

Sparse Matrix and Sparse Vector

MaxPlus.MP — Method

MP(A::SparseArray)

Convert a sparse array from classic algebra to a Max-Plus sparse array. By default, explicit Max-Plus zeros (ε, mp0, MP(-Inf)) are removed except if the parameter keepzeros is set to true.

Examples

julia> using SparseArrays

julia> A = MP(sparse([1, 2, 3], [1, 2, 3], [-Inf, 2, 0]))
3×3 Max-Plus sparse matrix with 2 stored entries:
  .   .   .
  .   2   .
  .   .   0

julia> A.nzval
2-element Max-Plus vector:
  2.0
  0.0

MP(I::AbstractVector, J::AbstractVector, V::AbstractVector)

Construct a sparse Max-Plus matrix such as S[I[k], J[k]] = V[k]

Examples

julia> MP([1; 2; 3], [1; 2; 3], [42; 43; 44])
3×3 Max-Plus sparse matrix with 3 stored entries:
  42    .    .
   .   43    .
   .    .   44

MP(V::SparseVector)

Convert a sparse vector from classic algebra to a Max-Plus sparse vector. By default, explicit Max-Plus zeros (ε, mp0, MP(-Inf)) are removed except if the parameter keepzeros is set to true.

Examples

julia> MP(sparse([1.0, 0.0, 1.0]))
3-element SparseVector{MP{Float64},Int64} with 2 stored entries:
  [1]  =  1.0
  [3]  =  1.0

julia> A = MP(sparse([1; -Inf; 3]))
3-element SparseVector{MP, Int64} with 2 stored entries:
  [1]  =  1.0
  [3]  =  3.0

MaxPlus.MP — Method

MP(A::SparseArray)

Convert a sparse array from classic algebra to a Max-Plus sparse array. By default, explicit Max-Plus zeros (ε, mp0, MP(-Inf)) are removed except if the parameter keepzeros is set to true.

Examples

julia> using SparseArrays

julia> A = MP(sparse([1, 2, 3], [1, 2, 3], [-Inf, 2, 0]))
3×3 Max-Plus sparse matrix with 2 stored entries:
  .   .   .
  .   2   .
  .   .   0

julia> A.nzval
2-element Max-Plus vector:
  2.0
  0.0

MP(I::AbstractVector, J::AbstractVector, V::AbstractVector)

Construct a sparse Max-Plus matrix such as S[I[k], J[k]] = V[k]

Examples

julia> MP([1; 2; 3], [1; 2; 3], [42; 43; 44])
3×3 Max-Plus sparse matrix with 3 stored entries:
  42    .    .
   .   43    .
   .    .   44

MP(V::SparseVector)

Convert a sparse vector from classic algebra to a Max-Plus sparse vector. By default, explicit Max-Plus zeros (ε, mp0, MP(-Inf)) are removed except if the parameter keepzeros is set to true.

Examples

julia> MP(sparse([1.0, 0.0, 1.0]))
3-element SparseVector{MP{Float64},Int64} with 2 stored entries:
  [1]  =  1.0
  [3]  =  1.0

julia> A = MP(sparse([1; -Inf; 3]))
3-element SparseVector{MP, Int64} with 2 stored entries:
  [1]  =  1.0
  [3]  =  3.0

MaxPlus.MP — Method

MP(I::AbstractVector, J::AbstractVector, V::AbstractVector)

Construct a sparse Max-Plus matrix such as S[I[k], J[k]] = V[k]

Examples

julia> MP([1; 2; 3], [1; 2; 3], [42; 43; 44])
3×3 Max-Plus sparse matrix with 3 stored entries:
  42    .    .
   .   43    .
   .    .   44

Unit Range

MaxPlus.MP — Method

MP(x::UnitRange)

Create a Max-Plus dense column vector from a given range.

Examples

julia> MP(1:3)
3-element Max-Plus Vector:
 1.0
 2.0
 3.0

MaxPlus.MP — Method

MP(x::StepRangeLen)

Create a Max-Plus dense column vector from a given range.

Examples

julia> MP(1.0:0.5:3.0)
5-element Max-Plus Vector:
  1.0
  1.5
  2.0
  2.5
  3.0

Overriden operators

Base.:+ — Method

+(x::MP, y::MP)

Max-Plus operator ⨁. Return the maximum of x and y as Max-Plus type. At least one parameter shall be a Max-Plus number (its conversion to Max-Plus is automatic).

Examples

julia> MP(1.0) + MP(3.0)
Max-Plus 3.0

julia> MP(1.0) + 3
Max-Plus 3.0

julia> 1 + MP(3.0)
Max-Plus 3.0

julia> MP(3.0) + -Inf
Max-Plus 3.0

Base.:* — Method

*(x::MP, y::MP)

Max-Plus operator ⨂. Return the sum of x and y as Max-Plus type. At least one parameter shall be a Max-Plus number (its conversion to Max-Plus is automatic).

Examples

julia> MP(1.0) * MP(3.0)
Max-Plus 4.0

julia> MP(1.0) * 3
Max-Plus 4.0

julia> 1 * MP(3.0)
Max-Plus 4.0

julia> MP(1.0) * -Inf
Max-Plus -Inf

Base.:\ — Method

\(A::ArrMP, b::ArrMP)

x = A \ b is a solution to A ⨂ x = b and its simply computed as inv(A) * b.

Examples

julia> A = [mp0 1 mp0; 2 mp0 mp0; mp0 mp0 3]
3×3 Max-Plus dense matrix:
  .   1   .
  2   .   .
  .   .   3

julia> B = [3 mp0 mp0; mp0 mp0 4; mp0 5 mp0]
3×3 Max-Plus dense matrix:
  3   .   .
  .   .   4
  .   5   .

julia> x = A \ B
3×3 Max-Plus dense matrix:
  .   .   2
  2   .   .
  .   2   .

julia> A * x == B
true

julia> A \ A
3×3 Max-Plus dense matrix:
  0   .   .
  .   0   .
  .   .   0

Missing docstring.

Missing docstring for Base.:(^)(::MP, ::Number). Check Documenter's build log for details.

Max-Plus constants

MaxPlus.mp0 — Constant

mp0

Create the constant Max-Plus zero (equals to -∞, minus infinity) in classic algebra which is the neutral for the ⨁ operator. See also mpzero.

Equivalent to ScicosLab code: %0 sugar notation for #(-%inf)

Examples

julia> mp0
Max-Plus -Inf

julia> mp0 * 5
Max-Plus -Inf

julia> mp0 + 5
Max-Plus 5.0

MaxPlus.ε — Constant

ε (\varepsilon)

Create the constant Max-Plus zero (equals to -∞, minus infinity) in classic algebra which is the neutral for the ⨁ operator. See also mpzero.

Equivalent to ScicosLab code: %0 sugar notation for #(-%inf)

Examples

julia> ε
Max-Plus -Inf

julia> ε * 5
Max-Plus -Inf

julia> ε + 5
Max-Plus 5.0

MaxPlus.mp1 — Constant

mp1

Create the constant Max-Plus one (equals to 0, zero) in classic algebra which is the neutral for operators ⨁ and ⨂. See also mpone.

Equivalent to ScicosLab code: %1 sugar notation for #(1)

Examples

julia> mp1
Max-Plus 0.0

julia> mp1 * 5
Max-Plus 5.0

julia> mp1 + 5
Max-Plus 5.0

MaxPlus.mpe — Constant

mpe

Create the constant Max-Plus one (equals to 0, zero) in classic algebra which is the neutral for operators ⨁ and ⨂. See also mpone.

Equivalent to ScicosLab code: %1 sugar notation for #(1)

Examples

julia> mpe
Max-Plus 0.0

julia> mpe * 5
Max-Plus 5.0

julia> mpe + 5
Max-Plus 5.0

MaxPlus.mptop — Constant

mptop

Create the constant Min-Plus one (equals to +∞, infinity) in classic algebra which is the neutral for operators ⨁ and ⨂.

Equivalent to ScicosLab code: %top = #(%inf)

Examples

julia> mptop
Max-Plus Inf

julia> mptop * 5
Max-Plus Inf

julia> mptop + 5
Max-Plus Inf

MaxPlus.mpI — Constant

mpI

Is the equivalent of LinearAlgebra.I but for Max-Plus type. Allow to create identity Max-Plus matrices.

Examples

julia> typeof(mpI)
LinearAlgebra.UniformScaling{MP}

julia> Matrix(mpI, 2, 2)
2×2 Max-Plus dense matrix:
   0.0   -Inf
  -Inf    0.0

julia> using LinearAlgebra

julia> Matrix(mpI, 2, 2) == Matrix{MP}(I, 2, 2)
true

Max-Plus Zeros

Base.zero — Method

zero(::MP)

Create the constant Max-Plus zero (equals to -∞, minus infinity) in classic algebra which is the neutral for the ⨁ operator. See also mp0 and ε.

Base.zero — Method

zero(::Type{MP})

Create the constant Max-Plus zero (equals to -∞, minus infinity) in classic algebra which is the neutral for the ⨁ operator. See also mp0 and ε.

MaxPlus.mpzero — Function

mpzero()

Create the constant Max-Plus zero (equals to -∞, minus infinity) in classic algebra which is the neutral for the ⨁ operator. See also mp0 and ε.

Examples

julia> mpzero()
Max-Plus -Inf

MaxPlus.mpzeros — Function

mpzeros(n::Int64)

Construct a Max-Plus zero n-by-m sparse matrix.

Examples

julia> mpzeros(2)
2-element SparseVector{MP,Int64} with 0 stored entries

mpzeros(n::Int64)

Construct a sparse Max-Plus zero m-by-n matrix.

Examples

julia> mpzeros(2,5)
2×5 SparseMatrixCSC{MP,Int64} with 0 stored entries
  .   .   .   .   .
  .   .   .   .   .

julia> full(mpzeros(2,5))
2×5 Max-Plus dense matrix:
  -Inf   -Inf   -Inf   -Inf   -Inf
  -Inf   -Inf   -Inf   -Inf   -Inf

mpzeros(A::Array{T})

Construct a sparse Max-Plus zero matrix of same dimension and of the same type that the matrix given as parameter.

Examples

julia> A=[1.0 2; 3 4]
2×2 Matrix{Float64}:
 1.0  2.0
 3.0  4.0

julia> mpzeros(A)
2×2 Max-Plus sparse matrix with 0 stored entries:
  .   .
  .   .

julia> mpzeros(MP(A))
2×2 Max-Plus sparse matrix with 0 stored entries:
  .   .
  .   .

Max-Plus Ones

Base.one — Method

one(::MP)

Create the constant Max-Plus one (equals to 0, zero) in classic algebra which is the neutral for operators ⨁ and ⨂. See also mp1 and mpe.

Base.one — Method

one(::MP)

Create the constant Max-Plus one (equals to 0, zero) in classic algebra which is the neutral for operators ⨁ and ⨂. See also mp1 and mpe.

MaxPlus.mpone — Function

mpone()

Create the constant Max-Plus one (equals to 0, zero) in classic algebra which is the neutral for operators ⨁ and ⨂. See also mp1 and mpe.

Examples

julia> mpone()
Max-Plus 0.0

MaxPlus.mpones — Function

mpones(n::Int64)

Construct a Max-Plus one n-by-1 matrix.

Examples

julia> mpones(2)
2-element Max-Plus vector:
  0.0
  0.0

mpones(m::Int64, n::Int64)

Construct a Max-Plus one m-by-n matrix.

Examples

julia> mpones(3,2)
3×2 Max-Plus dense matrix:
  0.0   0.0
  0.0   0.0
  0.0   0.0

mpones(A::Array{T})

Construct a sparse Max-Plus ones matrix of same dimension and of the same type that the matrix given as parameter.

Examples

julia> A=[1.0 2; 3 4]
2×2 Matrix{Float64}:
 1.0  2.0
 3.0  4.0

julia> mpones(A)
2×2 Max-Plus dense matrix:
  0.0   0.0
  0.0   0.0

Max-Plus Identity Matrix

MaxPlus.mpeye — Function

mpeye(n::Int64)

Construct a Max-Plus identity dense n-by-n matrix.

Examples

julia> mpeye(2)
2×2 Max-Plus dense matrix:
   0.0   -Inf
  -Inf    0.0

mpeye(m::Int64, n::Int64)

Construct a Max-Plus identity dense m-by-n matrix.

Examples

julia> mpeye(2, 3)
2×3 Max-Plus dense matrix:
   0.0   -Inf   -Inf
  -Inf    0.0   -Inf

mpeye(A::Array{T})

Construct a Max-Plus identity dense matrix of same dimension and of the same type that the matrix given as parameter.

Examples

julia> A=[1.0 2; 3 4]
2×2 Matrix{Float64}:
 1.0  2.0
 3.0  4.0

julia> mpeye(A)
2×2 Max-Plus dense matrix:
  0.0  -Inf
 -Inf   0.0

julia> mpeye(MP(A))
2×2 Max-Plus dense matrix:
  0.0  -Inf
 -Inf   0.0

Max-Plus Convertion

MaxPlus.plustimes — Function

plustimes(x::MP)

Convert a Max-Plus number to a number in standard algebra. An alternative way could be x.λ.

Examples

julia> plustimes(MP(1.0))
1.0

julia> typeof(plustimes(MP(1)))
Float64

plustimes(A::ArrMP)

Convert a Max-Plus dense matrix to a dense matrix in standard algebra.

Examples

julia> A=[MP(1.0) 2.0; ε mpe]
2×2 Max-Plus dense matrix:
   1.0   2.0
  -Inf   0.0

julia> plustimes(A)
2×2 Matrix{Float64}:
   1.0  2.0
  -Inf  0.0

plustimes(A::SpaMP)

Convert a Max-Plus sparse matrix to an sparse matrix in standard algebra.

Examples

julia> S = sparse(mpeye(2,2))
2×2 Max-Plus sparse matrix with 2 stored entries:
 0.0   ⋅
  ⋅   0.0

julia> findnz(S)
([1, 2], [1, 2], MP[0.0, 0.0])

julia> plustimes(S)
2×2 SparseMatrixCSC{Float64, Int64} with 2 stored entries:
 0.0   ⋅
  ⋅   0.0

julia> findnz(plustimes(S))
([1, 2], [1, 2], [0.0, 0.0])

MaxPlus.full — Function

full(::SpaMP})

Convert a sparse Max-Plus array to a dense Max-Plus array. Alternative function name: dense.

Examples

julia> full(mpzeros(2,5))
2×5 Max-Plus dense array:
 -Inf  -Inf  -Inf  -Inf  -Inf
 -Inf  -Inf  -Inf  -Inf  -Inf

MaxPlus.dense — Function

dense(::SpaMP})

Convert a sparse Max-Plus array to a dense Max-Plus array. Alternative function name: full.

Examples

julia> dense(mpzeros(2,5))
2×5 Max-Plus dense array:
 -Inf  -Inf  -Inf  -Inf  -Inf
 -Inf  -Inf  -Inf  -Inf  -Inf

MaxPlus.array — Function

array(::SpaMP})

Convert a sparse Max-Plus array to a dense non Max-Plus array.

Examples

julia> array(mpzeros(2,2))
2×2 Matrix{Float64}:
 -Inf  -Inf
 -Inf  -Inf

Power

Base.inv — Method

inv(A::ArrMP)

Return the inverse of a scalar or a Max-Plus matrix which is transpose(-A) if the matrix can be inversed. Throw an error if the matrix cannot be inversed.

Example 1: Scalar

julia> inv(MP(5))
Max-Plus -5.0

julia> MP(5)^-1
Max-Plus -5.0

Example 2: Square matrix inversible

julia> A = [mp0 1 mp0; 2 mp0 mp0; mp0 mp0 3]
3×3 Max-Plus dense matrix:
  .   1   .
  2   .   .
  .   .   3

julia> A^-1
3×3 Max-Plus transpose(dense matrix):
   .   -2    .
  -1    .    .
   .    .   -3

julia> inv(A)
3×3 Max-Plus transpose(dense matrix):
   .   -2    .
  -1    .    .
   .    .   -3

julia> A * inv(A)
3×3 Max-Plus dense matrix:
  0   .   .
  .   0   .
  .   .   0

julia> A * inv(A) == inv(A) * A == mpeye(A)
true

Example 3: Square matrix inversible

julia> A = [mp0 1 mp0; 2 mp0 mp0]
2×3 Max-Plus dense matrix:
  .   1   .
  2   .   .

julia> A^-1
3×2 Max-Plus transposed dense matrix:
   .   -2
  -1    .
   .    .

julia> inv(A)
3×2 Max-Plus transposed dense matrix:
   .   -2
  -1    .
   .    .

julia> A * inv(A)
2×2 Max-Plus dense matrix:
  0   .
  .   0

julia> inv(A) * A
3×3 Max-Plus dense matrix:
  0   .   .
  .   0   .
  .   .   .

julia> (A * inv(A)) != (inv(A) * A)
true

Example 4: Non inversible

julia> A = MP([1 2; 3 4])
2×2 Max-Plus dense matrix:
  1.0   2.0
  3.0   4.0

julia> inv(A)
ERROR: The matrix cannot be inversed

Max-Plus Matrix operations

MaxPlus.mptrace — Function

mptrace(A::Array)

Compute the trace of the matrix (summation of diagonal elements).

Examples

julia> mptrace([MP(1) 2; 3 4])
Max-Plus 4.0

MaxPlus.mpnorm — Function

mpnorm(A)

Compute the norm of the full or sparce matrix A. Return the largest entry minus smallest entry of A.

Examples

julia-repl
julia> A = MP([1 20 2;30 400 4;4 50 10])
3×3 Max-Plus dense matrix:
   1.0    20.0    2.0
  30.0   400.0    4.0
   4.0    50.0   10.0

julia> S = MP(sparse(A))
3×3 Max-Plus sparse matrix with 9 stored entries:
   1    20    2
  30   400    4
   4    50   10

julia> mpnorm(A)
Max-Plus 399

julia> mpnorm(S)
Max-Plus 399

MaxPlus.mpplus — Function

B = mpplus(A::ArrMP)

Compute A * A^* = A + A^2 + ... of a maxplus matrix A. See also mpstar.

Arguments

A : Max-Plus full square matrix.

B : Id + A + A^2 + ...

Examples

julia> mpplus(MP(1.0))
Max-Plus Inf

julia> mpplus(MP(-1.0))
Max-Plus -1

julia> A = MP([-3.0 -2; -1 0]); B = mpplus(A)
2×2 Max-Plus dense matrix:
  -3   -2
  -1    0

julia> B == (A * mpstar(A))
true

mpplus(x::MP)

Make x a 1x1 matrix then call mpplus(A::ArrMP). See mpplus(A::ArrMP) for more information.

MaxPlus.mpstar — Function

B = mpstar(A::ArrMP)

Solve x = Ax + I in the Max-Plus algebra. When there is no circuits with positive weight in G(A) (the incidence graph of A) B = I + A + ... + A^(n-1) where n denotes the order of the square matrix A.

See also mpplus.

Arguments

A : Max-Plus full square matrix.

B : Id + A + A^2 + ...

Examples

julia> mpstar(MP(1.0))
Max-Plus Inf

julia> mpstar(MP(-1.0))
Max-Plus 0

julia> mpstar(MP([1.0 2; 3 4]))
2×2 Max-Plus dense matrix:
  Inf   Inf
  Inf   Inf

julia> A = mpstar(MP([-3.0 -2;-1 0]))
2×2 Max-Plus dense matrix:
   0   -2
  -1    0

julia> B = mpstar(A)
2×2 Max-Plus dense matrix:
   0   -2
  -1    0

julia> B == (mpeye(2,2) + A)
true

julia> B == (B + A * A)
true

mpstar(x::MP)

Make x a 1x1 matrix then call mpstar(A::ArrMP). See mpstar(A::ArrMP) for more information.

MaxPlus.mpastarb — Function

mpastarb(A::ArrMP, b::ArrMP)

Max-Plus linear system solution.

Solve x = Ax + b in the Max-Plus algebra when there is no circuits with positive weight in G(A') (the incidence graph of A', that is it exists an arc from j to i if A_ij is nonzero).

TODO It is much more efficient in time and memory than mpstar(A) * b.

MaxPlus.howard — Function

howard(S::SpaMP)

TODO

Example 1: Irreducible matrix, only 1 eigenvalue.

julia> using SparseArrays

julia> S = sparse(MP([1 2; 3 4]))
2×2 Max-Plus Sparse Matrix with 4 stored entries:
  1   2
  3   4

julia> λ,v = howard(S)
(MP[4, 4], MP[2, 4])

# λ is constant
julia> (S * v) == (λ[1] * v)
true

Example 2: Two blocks diagonal matrix.

julia> using SparseArrays, LinearAlgebra

julia> S = sparse([mp0 2 mp0; mp1 mp0 mp0; mp0 mp0 2])
3×3 Max-Plus Sparse Matrix with 3 stored entries:
  .   2   .
  0   .   .
  .   .   2

julia> λ,v = howard(S)
(MP[1, 1, 2], MP[1, 0, 2])

# The entries of λ take two values
julia> (S / Matrix(Diagonal(λ))) * v == v
true

Example 3: Block triangular matrix with 2 eigenvalues.

julia> S = sparse([1 1; mp0 2])
2×2 Max-Plus sparse matrix with 3 stored entries:
  1   1
  .   2

julia> λ,v = howard(S)
(MP[2, 2], MP[1, 2])

julia> (S * v) == (λ[1] * v)
true

# But MP(1) is also eigen value
S * [0; %0] == MP(1) * [0; mp0]

Example 4: Block triangular matrix with 1 eigenvalue

julia> using SparseArrays, LinearAlgebra

julia> S = sparse([2 1; mp0 mp1])
2×2 Max-Plus sparse matrix with 3 stored entries:
  2   1
  .   0

julia> λ,v = howard(S)
(MP[2, 0], MP[2, 0])

# λ is not constant λ[1] is eigen value
# with eigen vector [v(1);0]
julia> (S / Matrix(Diagonal(λ))) * v == v
true

Display control of Max-Plus scalar and Matrices

MaxPlus.mp_change_display — Function

mp_change_display(style::Int)

Change the style of behavior of functions Base.show():

-Inf are displayed either with ε (style 2 or 3) or . symbols (style 1).

0 are displayed either with e(style 3) or '0' symbols (style 1 or 2).

else: -Inf and 0 are displayed in Julia default sytle (style 0).

If this function is not called, by default the ScicosLab style will be used (style 1).

MaxPlus.LaTeX — Function

LaTeX(S::MPSysLin)

Return the LaTeX code as string from the given Max-Plus linear system.

Examples

julia> 

LaTeX(io::IO, S::MPSysLin)

Display the LaTeX code from the given Max-Plus linear system.

Examples

julia> 

MaxPlus.mpshow — Function

mpshow(io::IO, S::MPSysLin)

Base function for displaying a Max-Plus linear system.

Examples

julia> 

Missing docstring.

Missing docstring for Base.show(::IO, ::MP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::ArrMP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::SpaMP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::MIME"text/plain", ::MP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::MIME"text/plain", ::ArrMP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::MIME"text/plain", ::SpaMP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::MIME"text/latex", ::MP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::MIME"text/latex", ::ArrMP). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::MIME"text/latex", ::SpaMP). Check Documenter's build log for details.

Index

MaxPlus.mp0

MaxPlus.mp1

MaxPlus.mpI

MaxPlus.mpe

MaxPlus.mptop

MaxPlus.ε

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

Base.:*

Base.:+

Base.:\

Base.inv

Base.one

Base.one

Base.zero

Base.zero

MaxPlus.LaTeX

MaxPlus.array

MaxPlus.dense

MaxPlus.full

MaxPlus.howard

MaxPlus.mp_change_display

MaxPlus.mpastarb

MaxPlus.mpeye

MaxPlus.mpnorm

MaxPlus.mpone

MaxPlus.mpones

MaxPlus.mpplus

MaxPlus.mpshow

MaxPlus.mpstar

MaxPlus.mptrace

MaxPlus.mpzero

MaxPlus.mpzeros

MaxPlus.plustimes