API: (min,+) Algebra

(min,+) constructor

Scalar Constructors

Missing docstring.

Missing docstring for MaxPlus.MI(::Float64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.MI(::Bool). Check Documenter's build log for details.

Dense Matrix and Dense Vector Constructors

Missing docstring.

Missing docstring for MaxPlus.MI(::Array). Check Documenter's build log for details.

Sparse Matrix Constructors

Missing docstring.

Missing docstring for MaxPlus.MI(::SparseMatrixCSC). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.MI(I::AbstractVector, J::AbstractVector, V::AbstractVector). Check Documenter's build log for details.

Sparse Vector Constructors

Missing docstring.

Missing docstring for MaxPlus.MI(::SparseVector). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.MI(I::AbstractVector, V::AbstractVector). Check Documenter's build log for details.

Unit Range Constructors

Missing docstring.

Missing docstring for MaxPlus.MI(::UnitRange). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.MI(::StepRangeLen). Check Documenter's build log for details.

Overloaded Algebraic Operators

Missing docstring.

Missing docstring for Base.zero(::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.zero(::Type{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.one(::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.one(::Type{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(+)(::MI, ::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(*)(::MI, ::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(^)(::MI, ::Number). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(/)(::MI, ::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(\)(::MI, ::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(-)(::MI, ::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:min(x::MI, y::MI). Check Documenter's build log for details.

(min,+) algebra to classic algebra conversion

Missing docstring.

Missing docstring for MaxPlus.plustimes(::MI). Check Documenter's build log for details.

(min,+) Constants

Missing docstring.

Missing docstring for MaxPlus.mi0. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.mi1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.mie. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.mitop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.miI. Check Documenter's build log for details.

(min,+) Dense matrices constructions

Missing docstring.

Missing docstring for Base.ones(MI, m::Int64, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.ones(MI, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.zeros(MI, m::Int64, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.zeros(MI, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.eye(MI, m::Int64, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.eye(MI, n::Int64). Check Documenter's build log for details.

(min,+) Sparse matrices constructions

Missing docstring.

Missing docstring for SparseArrays.spzeros(MI, m::Int64, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SparseArrays.spzeros(MI, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.speye(MI, m::Int64, n::Int64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.speye(MI, n::Int64). Check Documenter's build log for details.

(min,+) matrices Conversion

Missing docstring.

Missing docstring for MaxPlus.plustimes(A::Array{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.plustimes(S::SparseMatrixCSC{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.full(S::SparseMatrixCSC{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.dense(S::SparseMatrixCSC{MI}). Check Documenter's build log for details.

(min,+) Matrix operations

Missing docstring.

Missing docstring for Base.:(\)(::AbstractMatrix{MI}, ::AbstractMatrix{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(\)(::AbstractMatrix{MI}, ::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(\)(::MI, ::AbstractMatrix{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(/)(::AbstractMatrix{MI}, ::AbstractMatrix{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(/)(::AbstractMatrix{MI}, ::MI). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:(/)(::MI, ::AbstractMatrix{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.inv(::Matrix{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.star(A::Array{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.plus(A::Array{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.astarb(A::Array{MI}, b::Array{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.MIeigen. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.howard(S::SparseMatrixCSC{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.tr(A::Array{MI}). Check Documenter's build log for details.

LinearAlgebra.norm — Method

norm(A::Array{MP})

Compute the norm of the full or sparce matrix A. Return the largest entry minus smallest entry of A.

Examples

julia-repl
julia> using SparseArrays

julia> A = MP([1 20 2;30 400 4;4 50 10])
3×3 (max,+) dense matrix:
   1.0    20.0    2.0
  30.0   400.0    4.0
   4.0    50.0   10.0

julia> S = MP(sparse(A))
3×3 (max,+) sparse matrix with 9 stored entries:
   1    20    2
  30   400    4
   4    50   10

julia> mpnorm(A)
(max,+) 399

julia> mpnorm(S)
(max,+) 399

Display control of (min,+) scalar and Matrices

Missing docstring.

Missing docstring for MaxPlus.set_tropical_display. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPlus.LaTeX(io::IO, A::Matrix{MI}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for `#Base.show(::IO, ::Matrix{MI}) #Base.show(::IO, ::SparseMatrixCSC{MI})

Base.show(::IO, ::MIME"text/plain", A::Matrix{MI})`. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show(::IO, ::MIME"text/latex", A::Matrix{MI}). Check Documenter's build log for details.

Index

MaxPlus.mp0

MaxPlus.mp1

MaxPlus.mpI

MaxPlus.mpe

MaxPlus.mptop

MaxPlus.ε

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MP

MaxPlus.MPSysLin

Base.:*

Base.:*

Base.:+

Base.:+

Base.:-

Base.:/

Base.:/

Base.:/

Base.:/

Base.:/

Base.:==

Base.:\

Base.:\

Base.:\

Base.:\

Base.:^

Base.:|

Base.hcat

Base.inv

Base.min

Base.one

Base.one

Base.ones

Base.ones

Base.ones

Base.show

Base.show

Base.show

Base.show

Base.vcat

Base.zero

Base.zero

Base.zeros

Base.zeros

Base.zeros

LinearAlgebra.norm

LinearAlgebra.norm

LinearAlgebra.tr

MaxPlus.LaTeX

MaxPlus.LaTeX

MaxPlus.LaTeX

MaxPlus.astarb

MaxPlus.dense

MaxPlus.eye

MaxPlus.eye

MaxPlus.eye

MaxPlus.full

MaxPlus.howard

MaxPlus.mpeigen

MaxPlus.mpexplicit

MaxPlus.mpshow

MaxPlus.mpsimul

MaxPlus.plus

MaxPlus.plus

MaxPlus.plustimes

MaxPlus.plustimes

MaxPlus.plustimes

MaxPlus.set_tropical_display

MaxPlus.sparse_map

MaxPlus.speye

MaxPlus.speye

MaxPlus.spget

MaxPlus.star

MaxPlus.star

SparseArrays.spzeros

SparseArrays.spzeros

SparseArrays.spzeros